make pretty-printing work in rr
This fixes the "Attempt to write memory outside diversion session"
error that you would get when using the `pp` alias.

In order to preserve the correct replay behavior, rr must prevent the
state of the program from diverging from the recording. Thus, any gdb
command which writes to memory (or might do so), needs to be coerced
into writing to freshly-allocated memory which can then be thrown away
after the command is done. rr allocates a 'diversion area' where
writes can happen when we call a function, but it doesn't do this when
we merely try to set a variable (don't ask me why).