C++ Mode: Fontify correctly declarators with identifier preceded by &
The problem was bar in the following being spuriously recognised as a
function, and foo as a type, as though the & were a *: Foo foo (&bar);.

* lisp/progmodes/cc-engine.el (c-forward-decl-or-cast-1): New variable
got-function-name-prefix, which is set when an operator like * (but not &)
precedes the putative identifer in parentheses.  Test this variable when
deciding whether or not to "move the type backwards" to the previous
identifier.

* lisp/progmodes/cc-langs.el (c-type-decl-operator-prefix-key): New lang const
and var.